Move htings into __init__.py so automodule works
